data seed reafctor

+++ app/api/seed/route.ts
@@ -7,9 +7,9 @@ export async function POST(req: Request, res: Response) {
     const createdStudents = await prismaClient.data.createMany({
       data: seedData,
     });
-    return NextResponse.json(createdStudents);
+    return NextResponse.json(createdStudents, { status: 201 });
   } catch (error) {
-    console.error("Error seeding data:", error);
+    console.error("DATA_SEED_ERROR", error);
     return new Response("Error seeding data", { status: 500 });
   }
 }
@@ -20,10 +20,13 @@ export async function DELETE() {
     if (collegeDataCount === 0) {
       return new Response("No data to delete", { status: 400 });
     }
-    const deletedCollegeData = await prismaClient.data.deleteMany({});
-    return NextResponse.json(deletedCollegeData);
+
+    if (collegeDataCount > 0) {
+      const deletedCollegeData = await prismaClient.data.deleteMany({});
+      return NextResponse.json(deletedCollegeData, { status: 200 });
+    }
   } catch (error) {
-    console.error("Error deleting data:", error);
+    console.error("DATA_DELETE_ERROR", error);
     return new Response("Error deleting data", { status: 500 });
   }
 }
